Rename formation types to team-modes
